+++ frontend/src/components/Fallback.tsx
@@ -1,0 +1,5 @@
+export default function Fallback() {
+  return (
+    <h1 style={{ color: "red" }}>Loading中だよ...</h1>
+  )
+}
