+++ frontend/src/main.tsx
@@ -1,12 +1,10 @@
-import { StrictMode, lazy, Suspense } from 'react'
+import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 
-const App = lazy(() => import('./App.tsx'));
+import App from './App'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <Suspense fallback={<h1>Loading中だよ...</h1>}>
-      <App />
-    </Suspense>
+    <App />
   </StrictMode>,
 )
